+++ client/components/LogoutButton.test.tsx
@@ -5,35 +5,55 @@ import userEvent from "@testing-library/user-event";
 import renderer from "react-test-renderer";
 import LogoutButton from "./LogoutButton";
 
+const mockUseFetchContext = jest.fn();
+jest.mock("../context/FetchContext", () => ({
+    useFetchContext: () => {
+        return mockUseFetchContext();
+    },
+}));
+
 describe("LogoutButton", () => {
+    let fetchValues;
     beforeEach(() => {
-        Object.defineProperty(window, "sessionStorage", {
-            value: {
-                getItem: jest.fn(() => null),
-                removeItem: jest.fn(() => null),
+        fetchValues = {
+            state: {
+                token: "foo",
             },
-            writable: true,
-        });
+            action: {
+                clearToken: jest.fn(),
+            },
+        };
+        mockUseFetchContext.mockReturnValue(fetchValues);
     });
 
-    it("renders", () => {
-        const tree = renderer.create(<LogoutButton />).toJSON();
-        expect(tree).toMatchSnapshot();
+    it("renders correctly when logged in", async () => {
+        let tree;
+        await renderer.act(async () => {
+            tree = renderer.create(<LogoutButton />);
+        });
+        expect(tree.toJSON()).toMatchSnapshot();
+    });
+
+    it("renders correctly when logged out", async () => {
+        fetchValues.state.token = null;
+        let tree;
+        await renderer.act(async () => {
+            tree = renderer.create(<LogoutButton />);
+        });
+        expect(tree.toJSON()).toMatchSnapshot();
     });
 
     it("clears the token", async () => {
         render(<LogoutButton />);
+
+        expect(fetchValues.action.clearToken).not.toHaveBeenCalled();
 
         // Clicking a link will trigger a console error about navigation not being
         // implemented -- but that doesn't hurt the test, so let's suppress it by
         // mocking the error method.
         // TODO: figure out why and come up with a better solution than hiding the errors.
         jest.spyOn(console, "error").mockImplementation(jest.fn());
-
-        expect(sessionStorage.removeItem).not.toHaveBeenCalledWith("token");
-
         await userEvent.setup().click(screen.getByText("Log Out"));
-
-        expect(sessionStorage.removeItem).toHaveBeenCalledWith("token");
+        expect(fetchValues.action.clearToken).toHaveBeenCalled();
     });
 });